requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
mongoengine==0.23.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pydantic==1.8.1
pymongo==3.11.3
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dotenv==0.17.0
python-jose==3.2.0
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
